% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/actual_spouse.R
\name{actual_spouse}
\alias{actual_spouse}
\title{Returns the ID of an individual's spouse at a fixed date}
\usage{
actual_spouse(x = NULL, evdat = NULL, df_ped = NULL, df_fam = NULL)
}
\arguments{
\item{x}{An individual ID}

\item{evdat}{Dates of interest in the Year-month-day format (e.g. "1850-01-01")}

\item{df_ped}{A dataframe}

\item{df_fam}{A dataframe containing variables for date of union (\code{dat4}) and the IDs of the spouses (female: \code{idf}; male: code{idm})}
}
\description{
This function returns the ID of an individual's spouse at a fixed date (accounting for multiple marriages, if possible).
}
\examples{
df_ind <- get_exmpl_df()
df_ind$bdate <- sample(seq(as.Date("1774-12-31"), as.Date("1874-12-31"), 100), nrow(df_ind))
df_fam <- data.frame(idf = c(0,unique(df_ind$momid[df_ind$momid>0])), fall = "C")
actual_spouse(sample_kh(df_ind, df_fam),
   paste(sample(df_ind$bdate, 1)),
   df_ind, df_fam)
}
\keyword{kh}
\keyword{spouse}

