% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_kinstat.R
\name{add_kinstat}
\alias{add_kinstat}
\title{Add ID to maps}
\usage{
add_kinstat(id_list, evdat, kh_mat, kh_geo, kh_ind)
}
\arguments{
\item{id_list}{List of individual IDs of interests}

\item{kh_mat}{The event array}

\item{kh_ind}{Individual data}

\item{evdaz}{Date for estimatiom}

\item{list_geo}{A list of geo stuff.. with geopositional data for locations}
}
\description{
Returns a list of map objects, showing estimated locations for all living pedigree members for a given individual
\code{ggmap} provided plot of a map showing the estimated locations of living pedigree members for a given ID at a given date. The size, shape and color of symbols represent for a specific individual's age, sex, and relatedness to an indiviudal x.
}
\examples{
\dontrun{
plot_kinmap(40, "1860-01-01", df_ind, relate, evmat, geo$gmap, geo$df)
}
}
\keyword{kh}
\keyword{pedigree}

