% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_moms.R
\name{add_moms}
\alias{add_moms}
\title{Add iteratively IDs of mothers}
\usage{
add_moms(x, df_ind)
}
\arguments{
\item{x}{ID of individual for which a pedigree should be composed.}

\item{df_ind}{A dataframe for genealogical relations providing the variables 'id', 'momid' and 'dadid'}
}
\description{
This function returns a vector of IDs for a given individual matrilineal ancestry.
}
\examples{
\dontrun{
grap_kin(1067, df_ped)
}
}
\keyword{pedigree}

