% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as_date.R
\name{as_date}
\alias{as_date}
\title{Numeric Days from 1970 as Date}
\usage{
as_date(evdat)
}
\arguments{
\item{evdat}{A numerical to be converted to date}
}
\description{
This function returns a Date for a numeric value for the number of days since 1970-01-01.
}
\examples{
as_date(0)
}
\keyword{date}
\keyword{numeric}

