% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bioplot.R
\name{bio_plot}
\alias{bio_plot}
\title{Plot Biographies}
\usage{
bio_plot(id, df_ind, evmat)
}
\arguments{
\item{id}{An ID included in 'dimnames(evmat)'}

\item{df_ind}{Individual data frama}

\item{evmat}{An event array (see '?get_evmat')}
}
\description{
For a given ID sample, this function plots age intervals found in event array 'evmat'
}
\examples{
\dontrun{
valid_dates(1570, kh_mat)
}
}
\keyword{distance}
\keyword{pedigree}
\keyword{spatial}

