% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/built_kh.R
\name{built_kh}
\alias{built_kh}
\title{Built Krummhörn data files}
\usage{
built_kh(source_path = NULL, local_path = getwd(), file1 = "kindermc.sav",
  file2 = "familienmc.sav", silent = FALSE, n_sample = 10,
  force_new = FALSE, keep_files = NULL, set_seed = NULL)
}
\arguments{
\item{source_path}{Character string for the file path from which SPSS source(s) (e.g. `file1`) will be read.}

\item{local_path}{Character string for a file path to which 'kinlab' writes produced `.RData` files.}

\item{file1}{Character string for the mame of the first source file in SPSS format  (i.e. `.sav`).}

\item{file2}{An optional character string for the mame of the second source file in SPSS format  (i.e. `.sav`).}

\item{silent}{If TRUE, a final message giving the runtime of this function is omitted. (Default is FALSE)}

\item{n_sample}{Number of random draws, if n_sample < number of observations. (Default is 100)}

\item{force_new}{Should existing files be replaced? Default is FALSE}

\item{keep_files}{List specific files being kept in 'local_path' in case of `force_new=TRUE`.}

\item{set_seed}{Optional random seed being to allow reproducibility? Default is null, no seed is set.}
}
\description{
This function prepares SPSS data provided by the 'Krummhörn Family Reconstitution'
carried out by Voland et al. (e.g. Voland 2015) for being used with 'kinlab'.
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{pedigree}

