% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/built_ms.R
\name{built_ms}
\alias{built_ms}
\title{Prepare multi-state data}
\usage{
built_ms(id_sample = NULL, evmat = NULL, df_vars = NULL, min_n = 16,
  max_n = 1000)
}
\arguments{
\item{id_sample}{A sample of IDs (will be sample from, if length(id_sample)> max_n)}

\item{evmat}{An event matrix}

\item{df_vars}{Optional dataframe containing predictor variables}

\item{min_n}{Minimum number of cases per transtition to state under observation (Default to 16)}

\item{max_n}{Maximum number of total cases (Default to 1000)}
}
\description{
This function calls 'mstate::msprep()' with the appropriate argument for the transition matrix on an array of events (as returned by 'get_evmat()') by calling 'as.data.frame()'
returns a 'msdata' object (consisting of a table in long format and the coresponding transistion matrix)
}
\examples{
\dontrun{
df_ind <- get_exmpl_df()
df_fam <- data.frame(idf = c(0,unique(df_ind$momid[df_ind$momid>0])), fall = "C")
evmat <- get_evmat(df_ind, df_fam)
ms1 <- built_ms(evmat)
}
}
\keyword{distance}
\keyword{pedigree}
\keyword{spatial}

