% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/built_qb.R
\name{built_qb}
\alias{built_qb}
\title{Prepare data for pedigrees in Qu\enc{é}{e}bec}
\usage{

  built_qb(Individus = "/home/johow/Dropbox/qb/RPQA.MarcKlemp.individus.2012-01-27.sav",
  geo = "/home/johow/Dropbox/qb/geo.Rdata",
  location_list = "/home/johow/Dropbox/qb/qb_places.txt")
}
\arguments{
\item{Individus}{File path to  the 'Individus.sav' file (SPSS).}

\item{geo}{File path to  the 'geo.Rdata' file (R) containing geo data.}

\item{location_list}{File path to  the 'qb_places.txt' file (txt).}
}
\description{
This function returns a list of 10 elements:
\enumerate{
  \item{df_ind}{A dataframe with 80489 rows and 58 columns}
  \item{df_fam}{A dataframe with 34711 rows and 178 columns}
 \item{df_ped}{TBD}
 \item{relate}{TBD}
 \item{evmat}{TBD}
 \item{df_ev}{TBD}
 \item{evmat_bak}{TBD}
 \item{geo}{TBD}
 \item{evloc_list}{TBD}
 \item{evspc_list}{TBD}
}
}
\examples{
twinmat <- get_twinmat()
str(twinmat)
}
\keyword{import}
\keyword{kh}

