% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_gen.R
\name{count_gen}
\alias{count_gen}
\title{Count generation numbers}
\usage{
count_gen(df_ind = NULL)
}
\arguments{
\item{df_ind}{A dataframe with individual cases (id, dadid, momid).}
}
\description{
This function wraps the _pedigree::countGen()_ function to calculate generation number of each ID.
}
\examples{
df_ind <- get_exmpl_df()
count_gen(df_ind)
}
\keyword{import}
\keyword{kh}

