% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_kids8.R
\name{count_kids8}
\alias{count_kids8}
\title{Number of living children being younger than 8 years of a given individual at a given date}
\usage{
count_kids8(id, evdat, df_ped)
}
\arguments{
\item{id}{The ID of the individual parent (if any children exist) of interest (will be zero if no living children under age 8 alive)}

\item{evdat}{Date of interest in the Year-month-day format (e.g. "1850-01-01")}

\item{df_ped}{A dataframe}
}
\description{
This function returns  the number of living children being younger than 8 years of a given individual at a given date
}
\examples{
\dontrun{
count_kids8(1067, paste(as.Date(evmat["1067", 4, 1], origin = "1970-01-01")), df_ped)
}
}
\keyword{children}
\keyword{count}
\keyword{kh}

