% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/first_bdate.R
\name{first_bdate}
\alias{first_bdate}
\title{Get birth date of a mother's first born child}
\usage{
first_bdate(id, kh_mat)
}
\arguments{
\item{id}{Individual ID of parent(s)}

\item{kh_mat}{An event array (see  ?get_evmat )}
}
\description{
This function returns the birth date of a mother's first born child
}
\examples{
\dontrun{
df_ind$bdate <- sample(seq(as.Date("1774-12-31"), as.Date("1874-12-31"), 100), nrow(df_ind))
df_fam <- data.frame(idf = c(0,unique(df_ind$momid[df_ind$momid>0])), fall = "C")
sample_kh(df_ind, df_fam)
}
}
\keyword{pedigree}

