% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_age.R
\name{get_age}
\alias{get_age}
\title{Returns a given individual's age at a given date}
\usage{
get_age(x, evdat = NA, round.digits = NA, evmat)
}
\arguments{
\item{x}{Individual ID}

\item{evdat}{Dates of interest in the Year-month-day format (e.g. "1850-01-01")}

\item{round.digits}{Number of digits for rounding (default is NA, i.e. no rounding)}

\item{evmat}{A matrix for event data (see code{link{get_evmat}})}
}
\description{
This function returns the age of a given individual (with known birth date) at a given  date.
}
\examples{
\dontrun{
df_ind <- get_exmpl_df()
df_ind$bdate <- sample(seq(as.Date("1774-12-31"), as.Date("1874-12-31"), 100), nrow(df_ind))
df_fam <- data.frame(idf = c(0,unique(df_ind$momid[df_ind$momid>0])), fall = "C")
evmat <- get_evmat(df_ind, df_fam)
my_id <- sample_kh(df_ind, df_fam)
get_age(my_id, paste(df_ind$bdate[df_ind$id == my_id]+21*365-25), evmat)
}
}
\keyword{kh}
\keyword{spouse}

