% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_asfr.R
\name{get_asfr}
\alias{get_asfr}
\title{Age-specific fertility rates}
\usage{
get_asfr(x, df_ped, evmat, evmat_bak, include_twins = FALSE,
  treshold_date = as.Date("1830-01-01"))
}
\arguments{
\item{x}{The individual IDs of females of interests}

\item{df_ped}{A dataframe}

\item{evmat}{Event matrix (twins count as one)}

\item{evmat_bak}{Event matrix (including all twins)}

\item{include_twins}{Should twins count as multiple births? Default to FALSE}

\item{treshold_date}{Maximum date to include mothers with completed fertility. Default is "1830-01-01" for the Krummh\enc{ö}{oe}rn data.}
}
\value{
A dataframe with 'age', 'years' (i.e. number of years, where females have been exposed), 'nbirth' (number of births, or children being born see \code{include_twins}), 'nwomen' (number of women being ever exposed), 'asfr' (resulting age-specific fertility rate in births/1000 women/year).
}
\description{
This function computes age-specific fertility rates for a female, population being exposed (i.e. living years between age 15 and 45, see \href{http://www.un.org/esa/population/publications/WFD 2008/Metadata/ASFR.html}{United Nations, 2008}).
}
\examples{
\dontrun{
x <- which(attr(evmat, "dimnames")$id \%in\% df_ped$id[!is.na(df_ped$bdate) &
           df_ped$sex==2] &  (evmat[,20,1]- evmat[,1,1])/365.25 >= 15)
get_asfr(x, df_ped, evmat, evmat_bak)
}
}
\keyword{kh}
\keyword{pedigree}

