% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_brank.R
\name{get_brank}
\alias{get_brank}
\title{Individual birth rank}
\usage{
get_brank(x, df_ind = kh.full::kh_ind)
}
\arguments{
\item{x}{ID of interest}

\item{df_ind}{The individual data frame}
}
\description{
Returns an individuals birth rank (in case of twin births, it is not unique!).
}
\examples{
\dontrun{
get_brank(12)
}
}
\keyword{birth}
\keyword{rank}

