% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_brothers.R
\name{get_brothers}
\alias{get_brothers}
\title{Fraternal IDs for a given individual}
\usage{
get_brothers(x, df_ped)
}
\arguments{
\item{x}{ID of interest}

\item{df_ped}{A dataframe}
}
\description{
Returns the IDs of (maternal) brothers - if any exists - for a given individual.
}
\examples{
\dontrun{
get_brothers(1067, df_ped)
get_brothers(get_brothers(1067, df_ped)[1], df_ped)
}
}
\keyword{ID}
\keyword{brothers}
\keyword{kh}

