% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ceb.R
\name{get_ceb}
\alias{get_ceb}
\title{Children ever born}
\usage{
get_ceb(id, evmat)
}
\arguments{
\item{evmat}{An event matrix}

\item{id_sample}{A sample of IDs (will be sample from, if length(id_sample)> max_n)}
}
\description{
This function returns the number of children ever born for a given set of individuals and a given event matrix.
}
\examples{
\dontrun{
df_ind <- get_exmpl_df()
df_fam <- data.frame(idf = c(0,unique(df_ind$momid[df_ind$momid>0])), fall = "C")
evmat <- get_evmat(df_ind, df_fam)
ms1 <- built_ms(evmat)
}
}
\keyword{born}
\keyword{children}
\keyword{ever}
\keyword{pedigree}

