% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_dads.R
\name{get_dads}
\alias{get_dads}
\title{Paternal ID for a given individual}
\usage{
get_dads(x, df_ped)
}
\arguments{
\item{x}{An individual ID}

\item{df_ped}{A dataframe}
}
\description{
This function returns the ID of an individual's father.
}
\examples{
\dontrun{
get_dads(df_ped$id[2], df_ped)
}
}
\keyword{father}
\keyword{kh}

