% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Prepare data for pedigrees}
\usage{
get_data(kindermc = "/home/johow/Desktop/KH33_Final/kindermc.sav",
  familienmc = "/home/johow/Desktop/KH33_Final/familienmc.sav",
  geo = "/home/johow/Dropbox/variance/data/kh-geo.Rdata")
}
\arguments{
\item{kindermc}{File path to  the 'kindermc.sav' file (SPSS).}

\item{familienmc}{File path to  the 'familienmc.sav' file  (SPSS).}

\item{geo}{File path to  the 'kh-geo.Rdata' file (R).}
}
\description{
This function returns a list of 10 elements:
\enumerate{
  \item{df_kinder}{A dataframe with 80489 rows and 58 columns}
  \item{df_fam}{A dataframe with 34711 rows and 178 columns}
 \item{df_ped}{TBD}
 \item{relate}{TBD}
 \item{evmat}{TBD}
 \item{df_ev}{TBD}
 \item{evmat_bak}{TBD}
 \item{geo}{TBD}
 \item{evloc_list}{TBD}
 \item{evspc_list}{TBD}
}
}
\examples{
twinmat <- get_twinmat()
str(twinmat)
}
\keyword{import}
\keyword{kh}

