% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_df.R
\name{get_df}
\alias{get_df}
\title{Prepare data for pedigrees}
\usage{
get_df(df_ind = NULL, df_fam = NULL, silent = TRUE, get_fam = FALSE,
  replace_sex = TRUE)
}
\arguments{
\item{df_ind}{A dataframe with individual cases}

\item{df_fam}{A dataframe with family cases}

\item{silent}{Should messages be suppressed (Default is FALSE)}

\item{get_fam}{If TRUE, the dataframe 'df_fam with additional columns 'idm' and 'idf' is returned instead (default is FALSE)}

\item{replace_sex}{Should entries be corrected for individuals appearing as mothers or fathers? (default is TRUE)}
}
\description{
This function returns a dataframe being required to contruct pedigrees via the _kinship2::pedigree()_ function
}
\examples{
\dontrun{
kh_df <- get_df(df_ind, df_fam, get_fam=FALSE)
}
}
\keyword{import}
\keyword{kh}

