% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_df_ev_qb.R
\name{get_df_ev_qb}
\alias{get_df_ev_qb}
\title{A dataframe for individual events}
\usage{
get_df_ev_qb(df_ped = qb[{     "df_ped" }])
}
\arguments{
\item{df_ped}{A dataframe containing as columns "id", "bdate", "bplace", "gebkk", "ddate", "dplace", and "todkk"}

\item{df_fam}{A dataframe containing as columns "idm","idf", "dat4", "cort", and "fall"}
}
\description{
This function returns a dataframe (in "long format", i.e. an individual's ID spanning multiple rows) for event data (date, location, and type of event).
}
\examples{
get_dads(kh[["df_ped"]]$id[2])
 get_dads(get_kids(get_dads(kh[["df_ped"]]$id[2])))
}
\keyword{events}
\keyword{kh}

