% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_dist}
\alias{get_dist}
\title{Spatial distance between individuals at a specific date}
\usage{
get_dist(id, id2, evdat, evmat, map_dist)
}
\arguments{
\item{id}{ID of individual for which a pedigree should be composed.}

\item{id2}{ID of individual for which distance should be computed}

\item{evdat}{A specific date in the format ``Year-month-day''}

\item{evmat}{Event matrix}

\item{dist_mat}{A distance matrix}
}
\description{
This function returns the spatial distance in kilometres between individual x and pedigree members at a specific date
}
\examples{
\dontrun{
attach(kh33::kh)
get_dist(40, grap_ped(40, df_ped), "1850-01-01", dist_mat = geo[["map_dist"]], evmat=evmat)
detach(kh33::kh)
}
}
\keyword{distance}
\keyword{pedigree}
\keyword{spatial}

