% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_dt_ev.R
\name{get_dt_ev}
\alias{get_dt_ev}
\title{A datatable for individual events}
\usage{
get_dt_ev(df_ind = NULL, df_fam = NULL)
}
\arguments{
\item{df_ind}{A dataframe containing as columns "id", "bdate", "bplace", "bstatus", "ddate", "dplace", and "dstatus"}

\item{df_fam}{A dataframe containing as columns "idm","idf", "dat4", "cort", and "fall"}
}
\description{
This function returns a data.table::data.table object (in "long format", i.e. an individual's ID spanning multiple rows) for event data (date, location, and type of event).
}
\examples{
\dontrun{
df_ind <- get_exmpl_df()
df_fam <- data.table::data.table(idf = c(0,unique(df_ind$momid[df_ind$momid>0])), fall = "C")
df_ev <- get_df_ev(df_ind, df_fam)
}
}
\keyword{events}
\keyword{kh}

