% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_evmat2.R
\name{get_evmat2}
\alias{get_evmat2}
\title{A matrix for event data}
\usage{
get_evmat2(df_ev = NULL, vars = list(evdat = "evdat", evid = "evid", evloc =
  "evloc", evspc = "evspc", status = "status", evtyp = "evtyp", id = "id"),
  events = list(birth = "*", child = "#", death = "+"), censored = FALSE)
}
\arguments{
\item{df_ev}{A dataframe in long format containing individual events}

\item{vars}{A list of variables for chronological date, ID related to event, location, special, status, and type of event)}

\item{events}{A list of labels used in \code{vars[["evtyp"]]} to indicate births, born children, and deaths}

\item{censored}{A logical indicator, whether censored dates should be included (Default is FALSE)}
}
\description{
This function returns a matrix describing event data in the form date, id, location, special, status. Here, gemini births count as one single event (see get_evmat_twins() for including the birth of twins as multiple events).
}
\examples{
\dontrun{
df_fam <- data.frame(idf = c(0,unique(df_ind$momid[df_ind$momid>0])), fall = "C")
evmat <- get_evmat(df_ind, df_fam)
}
}
\keyword{events}
\keyword{kh}
\keyword{kin}

