% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_evmat_bak.R
\name{get_evmat_bak}
\alias{get_evmat_bak}
\title{A matrix for event data}
\usage{
get_evmat_bak(df_ped = kh[["df_ped"]], df_fam = kh[["df_fam"]])
}
\arguments{
\item{df_ped}{A dataframe containing genealogical information}

\item{df_fam}{A dataframe of the Krummh\enc{ö}{oe}rn family data}
}
\description{
This function returns a matrix describing event data in the form date, id, location, special, status. Here, gemini births are included as multiple events (with each cotwin being born at the same date).
}
\keyword{events}
\keyword{kh}
\keyword{kin}

