% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_evmat_qb.R
\name{get_evmat_qb}
\alias{get_evmat_qb}
\title{A matrix for event data in Qu\enc{é}{e}bec}
\usage{
get_evmat_qb(df_ped = qb[["df_ped"]])
}
\arguments{
\item{df_ped}{A dataframe containing genealogical information}
}
\description{
This function returns a matrix describing event data in the form date, id, location, special, status. Here, gemini births count as one single event (see get_evmat_twins() for including the birth of twins as multiple events).
}
\keyword{events}
\keyword{kin}
\keyword{qb}

