% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_exmpl.R
\name{get_exmpl_df}
\alias{get_exmpl_df}
\title{Small pedigree dataframe for example}
\usage{
get_exmpl_df()
}
\description{
This function returns an example pedigree dataframe (no arguments).
}
\examples{
\dontrun{
get_exmpl_df()
}
}
\keyword{pedigree}

