% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_file_path.R
\name{get_file_path}
\alias{get_file_path}
\title{Source and local file paths}
\usage{
get_file_path(path_type = c("local", "source", "inst", "cloud")[1],
  db = c("kh", "qb")[1], dir = "data", sub = NULL,
  linux_source = paste0("/home/johow/Dropbox/db/", db),
  win_source = paste0("C:/Users/JJ/Desktop/db/", db),
  inst_source = ifelse(grepl("linux", R.version["platform"]),
  "/media/johow/USB DISK/inst/data", "H://inst/data"), pkg = "kinlab",
  my_lib = NULL)
}
\arguments{
\item{path_type}{Either 'local' for the 'kinlab' folder installed on the current machine, or 'source' folders specified in `linux_source` and `win_source`, or 'cloud' for predefined Dropbox folder)}

\item{db}{A name of a data source, currently either 'kh' (i.e. Krummhörn, as default) or 'qb' (Quebec).}

\item{dir}{Name of the local kinlab directory in library (Default is 'inst')}

\item{sub}{NULL or name of subdirectory (if required, default is `file.path("data", db))}

\item{linux_source}{Linux source folder (including subdirectory for `db`)}

\item{win_source}{Windows source folder (including subdirectory for `db`)}

\item{inst_source}{Source folder for library directory}

\item{pkg}{Internal use (not to be changed by user)}

\item{my_lib}{Optionally allows setting a library directory manually}
}
\description{
Returns file path to directories containing `kinlab` sources
and local data on linux and windows machines.
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{file}

