% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_foundfam_id.R
\name{get_foundfam_id}
\alias{get_foundfam_id}
\title{IDs of an individual's founded family's members}
\usage{
get_foundfam_id(x, df_ped)
}
\arguments{
\item{x}{An individual ID}

\item{df_ped}{A dataframe for genealogical relations providing the variables 'id', 'momid' and 'dadid'}
}
\description{
For a given individual, this function returns IDs of the parents and their (shared) offspring being ever born.
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{kh}
\keyword{pedigree}

