% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_geo.R
\name{get_geo}
\alias{get_geo}
\title{Geopositional Data List}
\usage{
get_geo(data_source = "kh")
}
\arguments{
\item{data_source}{A character string for the name of a given population (Currently "kh" for 'Krummhörn' is the only valid value and default)}
}
\description{
This function returns a list of 3 elements:
\enumerate{
  \item{df_geo}{A dataframe with 33 rows for the Krummhörn Kirchspiele's Longitude and Lattitude (as returned from Google Maps)}
  \item{maps}{A list of 2 Maps returned from 'ggmap::ggmap()'}
 \item{mat}{A distance Matrix for all 33 Kirchspiele, as returned from Google Maps}
}
}
\examples{
twinmat <- get_twinmat()
str(twinmat)
}
\keyword{import}
\keyword{kh}

