% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_grasen.R
\name{get_grasen}
\alias{get_grasen}
\title{Get family's grasen value for a given ID of a parent}
\usage{
get_grasen(x = NULL, df_fam = NULL)
}
\arguments{
\item{x}{ID of a parent}

\item{df_fam}{A dataframe containing ID and grasen values}
}
\description{
This function returns a family's value of 'grasen' (a measure of land size) for a given ID of a parent.
}
\examples{
\dontrun{
df_ind$bdate <- sample(seq(as.Date("1774-12-31"), as.Date("1874-12-31"), 100), nrow(df_ind))
df_fam <- data.frame(idf = c(0,unique(df_ind$momid[df_ind$momid>0])), fall = "C")
sample_kh(df_ind, df_fam)
}
}
\keyword{pedigree}

