% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_kh.R
\name{get_kh}
\alias{get_kh}
\title{Wrapper function to import and format Krummh\enc{ö}{oe}rn data for children and families from SPSS files}
\usage{
get_kh(kindermc = "", familienmc = "")
}
\arguments{
\item{kindermc}{File path to  the 'kindermc.sav' SPSS file.}

\item{familienmc}{File path to  the 'familienmc.sav' SPSS file}
}
\description{
This function returns a list containing three data.frames:
\enumerate{
  \item Imported 'kindermc.sav' SPSS file as read from source.
  \item Imported 'familienmc.sav' SPSS file as read from source.
  \item Dataframe holding individual IDs and those of the parents.
}
}
\examples{
twinmat <- get_twinmat()
str(twinmat)
}
\keyword{import}
\keyword{kh}

