% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_kids.R
\name{get_kids}
\alias{get_kids}
\title{Returns the ID(s) of an individual's child(ren)}
\usage{
get_kids(x, df_ped)
}
\arguments{
\item{x}{An individual ID}

\item{df_ped}{A dataframe}
}
\description{
This function returns the ID(s) of an individual's children
}
\examples{
\dontrun{
get_kids(df_ped$id[2], df_ped)
detach(kh33::kh)
}
}
\keyword{children}
\keyword{kh}

