% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_kindata.R
\name{get_kindata}
\alias{get_kindata}
\title{A dataframe describing the kin network for a given ID at a given date.}
\usage{
get_kindata(id, evdat, df_ped = kh[["df_ped"]],
  map_dist = kh[["geo"]][["map_dist"]], kin.only = TRUE)
}
\arguments{
\item{id}{ID of individual for which specific kin network should be compiled}

\item{evdat}{A specific date of interest}

\item{df_ped}{A dataframe for genealogical relations providing the variables 'id', 'momid' and 'dadid'}

\item{map_dist}{A distance matrix, which contains locations of individuals}

\item{kin.only}{If FALSE, also kin being genetically unrelated both to the given ID and his or her spouse(s) will be included (Default is TRUE)}
}
\description{
This function returns a dataframe describing kin networks at specific dates documented for one or more given IDs
}
\examples{
get_kindata(1067, "1801-01-01")
}
\keyword{events}
\keyword{kh}
\keyword{kin}

