% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_kinev.R
\name{get_kinev}
\alias{get_kinev}
\title{Pedigree members with documented event data}
\usage{
get_kinev(x, df_ped, evmat, ped = NULL)
}
\arguments{
\item{x}{An individual ID}

\item{df_ped}{A dataframe}

\item{evmat}{A matrix providing event data}

\item{ped}{A pedigree object}
}
\description{
This function returns the IDs for those pedigree members of a given individual for which event data is available.
}
\examples{
\dontrun{
##
}
}
\keyword{events}
\keyword{kh}
\keyword{pedigree}

