% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_kinloc.R
\name{get_kinloc}
\alias{get_kinloc}
\title{Returns the estimated locality of a given individual at a fixed date}
\usage{
get_kinloc(id, evdat, evmat)
}
\arguments{
\item{id}{An individual ID}

\item{evdat}{Date of interest in the Year-month-day format (e.g. "1850-01-01")}

\item{evmat}{Event matrix}
}
\description{
This function returns the estimated locality of a given individual at a given date. Locality is estimated from the individual's last record before or at the date of interest.
}
\examples{
get_min_birthage(1067, kh33::kh$evmat_bak)
}
\keyword{kh}
\keyword{locality}

