% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_kinset.R
\name{get_kinset}
\alias{get_kinset}
\title{A snapshot of an individual's kinship structure at a fixed date}
\usage{
get_kinset(x, evdat, ped, df_ind, df_fam, evmat, map_dist)
}
\arguments{
\item{x}{ID of individual for which a pedigree should be composed.}

\item{evdat}{A specific date in the format ``Year-month-day''}

\item{ped}{Pedigree of `x`}

\item{df_ind}{Dataframe for individuals}

\item{df_fam}{Dataframe for unions}

\item{evmat}{Event matrix}

\item{map_dist}{A distance matrix}
}
\description{
This function returns a dataframe containing pedigree members of individual `x` being alive at date `evdat`.
}
\examples{
\dontrun{
attach(kh33::kh)
get_dist(40, grap_ped(40, df_ped), "1850-01-01", dist_mat = geo[["map_dist"]], evmat=evmat)
detach(kh33::kh)
}
}
\keyword{distance}
\keyword{pedigree}
\keyword{spatial}

