% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_min_birthage.R
\name{get_min_birthage}
\alias{get_min_birthage}
\title{Returns the minimum age at birth
This function returns a mother's minimum age at birth}
\usage{
get_min_birthage(mother.id = NULL, evmat_bak = NULL)
}
\arguments{
\item{mother.id}{An individual mother ID}

\item{evmat_bak}{An event matrix}
}
\description{
Returns the minimum age at birth
This function returns a mother's minimum age at birth
}
\examples{
\dontrun{
get_min_birthage(1067, kh33::kh$evmat_bak)
}
}
\keyword{kh}
\keyword{spouse}

