% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_min_twinage.R
\name{get_min_twinage}
\alias{get_min_twinage}
\title{Returns a mother's minimum age at birth, considering only twin births}
\usage{
get_min_twinage(twinmother.id = NULL, evmat_bak = NULL)
}
\arguments{
\item{twinmother.id}{Dates of interest in the Year-month-day format (e.g. "1850-01-01")}

\item{evmat_bak}{An event matrix}
}
\description{
This function returns a  mother's minimum age at birth, considering only twin births
}
\examples{
\dontrun{
get_min_birthage(1067, kh33::kh$evmat_bak)
}
}
\keyword{kh}
\keyword{spouse}

