% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_moms.R
\name{get_moms}
\alias{get_moms}
\title{Returns the ID of an individual's mother}
\usage{
get_moms(x, df_ped)
}
\arguments{
\item{x}{An individual ID}

\item{df_ped}{A dataframe}
}
\description{
This function returns the ID of an individual's mother
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{kh}
\keyword{mother}

