% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_qb.R
\name{get_qb}
\alias{get_qb}
\title{Wrapper function to import and format Qu\enc{é}{e}bec data for children and families from SPSS files}
\usage{

  get_qb(Individus = "/home/johow/Dropbox/qb/RPQA.MarcKlemp.individus.2012-01-27.sav")
}
\arguments{
\item{Individus}{File path to  the 'Individus.sav' SPSS file.}
}
\description{
This function returns a list containing three data.frames:
\enumerate{
  \item Imported 'Individus' SPSS file as read from source.
  \item Dataframe holding individual IDs and those of the parents.
  \item Matrix describing event data in the form date, id, location, special, status. Here, gemini births count as one single event (see get_evmat_twins() for including the birth of twins as multiple events).
  \item Matrix describing event data in the form date, id, location, special, status. Here, gemini births count as one single event (see get_evmat_twins() for including the birth of twins as multiple events).
  \item Matrix describing twin relations in the form 'id1-id2-code', where code stands for 1=monozygotous, 2=dizygotous, 3=unknown. Here, zygosity is soley based on sex and therefore only twins of different sex are given code=2 (i.e. dizygotous), while same-sex twins are classified as unknown (code=3). In the case of triplets (and twins of higher order), each individual will have relations to each of his co-twins.
  }
}
\examples{
notrun{
twinmat <- get_twinmat()
# str(twinmat)
}
}
\keyword{import}
\keyword{qb}

