% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_resid_rv.R
\name{get_resid_rv}
\alias{get_resid_rv}
\title{Average age-specific residual reproductive value}
\usage{
get_resid_rv(id, df_ped, evmat, evmat_bak, age = NULL)
}
\arguments{
\item{id}{The individual IDs of females of interests}

\item{df_ped}{A dataframe for genealogical relations providing the variables 'id', 'momid' and 'dadid'}

\item{evmat}{Event matrix}

\item{evmat_bak}{An event matrix}

\item{age}{Age for which the residual reproductive value should be computed.}
}
\value{
The average number of offspring, which can be expected for individuals in \code{id} at age \code{age}.
}
\description{
This function computes residual reproductive values based on estimated survival probabilites and age-specific fertiliy rates (see \href{http://en.wikipedia.org/wiki/Life_history_theory#Reproductive_value_and_costs_of_reproduction}{Wikipedia, 2015}).
}
\examples{
\dontrun{
get_resid_rv(sample(df_ped$momid[df_ped$momid>0], 100), df_ped, age=30)
}
}
\keyword{kh}
\keyword{pedigree}

