% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_sex.R
\name{get_sex}
\alias{get_sex}
\title{Return the sex of a given ID}
\usage{
get_sex(x, df_ped)
}
\arguments{
\item{x}{Ab individual ID}

\item{df_ped}{A dataframe}
}
\description{
This function returns the sex of a given individual.
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{kh}
\keyword{sex}

