% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_spouses.R
\name{get_spouses}
\alias{get_spouses}
\title{Return spousal ID(s) for a married individual}
\usage{
get_spouses(x, df_ped)
}
\arguments{
\item{x}{ID of married individual, whose spouses should be returned}

\item{df_ped}{A dataframe}
}
\description{
This function returns individual IDs of a given individual spouses (if possible, i.e. individual must be a parent).
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{IDs}
\keyword{spouses}

