% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_twin_id.R
\name{get_twin_id}
\alias{get_twin_id}
\title{Returns the IDs of twins of a twin mother}
\usage{
get_twin_id(twinmother.id = NULL, evmat_bak = NULL, relate = NULL)
}
\arguments{
\item{twinmother.id}{Dates of interest in the Year-month-day format (e.g. "1850-01-01")}

\item{evmat_bak}{An event matrix}

\item{relate}{A twin relation matrix}
}
\description{
This function returns the IDs of twins of a twin mother
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{kh}
\keyword{spouse}

