% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_twinmat.R
\name{get_twinmat}
\alias{get_twinmat}
\title{Wrapper function to return a matrix for twin relations}
\usage{
get_twinmat(df_ped = NULL, bspecial_exclude = NULL, bstatus_ok = NULL)
}
\arguments{
\item{df_ped}{A dataframe containing the children's birth dates and and the ID of their parents}

\item{bspecial_exclude}{Possible values of 'bspecial' to exclude as twin births}

\item{bstatus_ok}{A value of the variabe 'bstatus', which indicates that a birth date is adequately documented. )}
}
\description{
This function returns a matrix describing twin relations in the form 'id1-id2-code', where code stands for 1=monozygotous, 2=dizygotous, 3=unknown. Here, zygosity is soley based on sex and therefore only twins of different sex are given code=2 (i.e. dizygotous), while same-sex twins are classified as unknown (code=3). In the case of triplets (and twins of higher order), each individual will have relations to each of his co-twins.
}
\examples{
\dontrun{
twinmat <- get_twinmat(df_ped, df_kinder)
}
}
\keyword{pedigree}

