% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_twinmat_qb.R
\name{get_twinmat_qb}
\alias{get_twinmat_qb}
\title{Wrapper function to return a matrix for twin relations}
\usage{
get_twinmat_qb(df_ped = qb[["df_ped"]])
}
\arguments{
\item{df_ped}{A dataframe containing the children's birth dates and and the ID of their parents}

\item{df_kinder}{A dataframe containing information on the type of certificate (required to exclude siblings having their baptism on the same day from being classified as twins)}
}
\description{
This function returns a matrix describing twin relations in the form 'id1-id2-code', where code stands for 1=monozygotous, 2=dizygotous, 3=unknown. Here, zygosity is soley based on sex and therefore only twins of different sex are given code=2 (i.e. dizygotous), while same-sex twins are classified as unknown (code=3). In the case of triplets (and twins of higher order), each individual will have relations to each of his co-twins.
}
\examples{
twinmat <- get_twinmat()
str(twinmat)
}
\keyword{pedigree}

