% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grap_df.R
\name{grap_df}
\alias{grap_df}
\title{A pedigree dataframe for subset of limited size}
\usage{
grap_df(x = NULL, df_ind = NULL, subsample_id = NULL, min_size = 300,
  max_count = 9)
}
\arguments{
\item{x}{An individual ID}

\item{df_ind}{A dataframe for genealogical relations providing the variables 'id', 'momid' and 'dadid'}

\item{min_size}{Minimum size, i.e. number of family members}

\item{max_count}{Maximum number of iterations}
}
\description{
For a given individual, this function returns a pedigree dataframe for close kin
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{kh}
\keyword{pedigree}

