% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grap_kin.R
\name{grap_kin}
\alias{grap_kin}
\title{Add iteratively IDs of related family members}
\usage{
grap_kin(x, df_ind, max_size, max_cycles = 8)
}
\arguments{
\item{x}{ID of individual for which a pedigree should be composed.}

\item{df_ind}{A dataframe for genealogical relations providing the variables 'id', 'momid' and 'dadid'}

\item{max_size}{Maximum size of pedigree of interest (Default is 300).}

\item{max_cycles}{Maximum number of iterations (adding further relatives).}
}
\description{
This function returns a vector of family members IDs for a given individual by iteratively adding IDs of offspring and parents and vice versa for a given times.
}
\examples{
\dontrun{
grap_kin(1067, df_ped)
}
}
\keyword{pedigree}

