% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grap_ped.R
\name{grap_ped}
\alias{grap_ped}
\title{Return a compact, individual-centered pedigree by iteratively adding relatives until a given minimum pedigree size is reached.}
\usage{
grap_ped(x = NULL, df_ind = NULL, relate = NULL, subsample_id = NULL,
  min_size = 300, max_count = 9)
}
\arguments{
\item{x}{ID of individual for which a pedigree should be composed.}

\item{df_ind}{A dataframe for genealogical relations providing the variables 'id', 'momid' and 'dadid'}

\item{relate}{A relationshi matrix specifying twin zygosity (1=MZ, 2=DZ, 3=unknown)}

\item{subsample_id}{A vector of IDs (Set 'as.numeric(dimnames(evmat)$id') for IDs with events)}

\item{min_size}{Minimum size of pedigree of interest (Default is 200)}

\item{max_count}{Maximum number of iterations in adding relatives to pedigree}
}
\description{
This function returns a pedigree object (as specified in the kinship2-package) for a given individual by iteratively adding spouses and first degree relatives until a given size is reached.
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{pedigree}

