% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_alive.R
\name{is_alive}
\alias{is_alive}
\title{Test for estimated survival status of a given individual at a given date.}
\usage{
is_alive(x, evdat, df_ind)
}
\arguments{
\item{x}{An individual's ID}

\item{evdat}{Date of interest for estimated survival status of individual x in the format "Year-month-day".}

\item{df_ind}{A dataframe containing birth and death dates of individuals in x}
}
\description{
Returns estimated survival status of a given individual at a given date. Estimates are based on recorded deaths and the assumption that individuals neither dying nor being mentioned elsewhere did survive at least until the age of 15 (and later possibly outmigrated).
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{status}
\keyword{survival}

