% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_related.R
\name{is_related}
\alias{is_related}
\title{Test for relatedness}
\usage{
is_related(id1, id2, ped = NA, df_ped, relate)
}
\arguments{
\item{id1}{ID of first individual}

\item{id2}{ID of second individual}

\item{ped}{A pedigree object containing both individuals}

\item{df_ped}{A dataframe}

\item{relate}{A realtionship matrix for twin zygosity}
}
\description{
This function tests for relatedness by computing the relatedness coefficient between two given individuals (id1 and id2) in a given pedigree (ped). It returns either 1 for 'related' (i.e. r > 0) or 0 for 'unrelated' (r = 0).
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{pedigree}
\keyword{relatedness}

