% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kin_crawler.R
\name{kin_crawler}
\alias{kin_crawler}
\title{Add IDs of family members}
\usage{
kin_crawler(x, df_ped)
}
\arguments{
\item{x}{ID of individual for which a pedigree should be composed.}

\item{df_ped}{A dataframe for genealogical relations providing the variables 'id', 'momid' and 'dadid'}
}
\description{
This function returns a vector of family members IDs for a given individual by iteratively adding IDs of offspring and parents and vice versa for a given times.
}
\examples{
\dontrun{
grap_kin(1067, df_ped)
}
}
\keyword{pedigree}

