% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_seq.R
\name{map_seq}
\alias{map_seq}
\title{Map sequence}
\usage{
map_seq(my_id = NULL, kin_list = NULL, my_map = NULL, geo_df = NULL,
  return_gg = TRUE, map_extent = "device", scale_factor = 1,
  limit_n = NULL)
}
\arguments{
\item{my_id}{The individual ID of interests}

\item{kin_list}{A nested list of dataframes providing the variables 'id', 'sex', 'relmom', 'reldad', 'age', and 'kinloc'}

\item{my_map}{A ggmap object}

\item{geo_df}{A dataframe with geopositional data for locations}

\item{return_gg}{A logical indicator whether the plots should be returned (if TRUE) or displayed (if  FALSE)}

\item{map_extent}{How much of the plot should the map take up? (default is "device")}

\item{scale_factor}{Ratio of scaling in relation to 1 (Default).}
}
\description{
Returns a list of map objects, showing estimated locations for all living pedigree members for a given individual
\code{ggmap} provided plot of a map showing the estimated locations of living pedigree members for a given ID at a given date. The size, shape and color of symbols represent for a specific individual's age, sex, and relatedness to an indiviudal x.
}
\examples{
\dontrun{
plot_kinmap(40, "1860-01-01", df_ind, relate, evmat, geo$gmap, geo$df)
}
}
\keyword{kh}
\keyword{pedigree}

