% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mat2bio.R
\name{mat2bio}
\alias{mat2bio}
\title{Converts event array to Biograph}
\usage{
mat2bio(evmat, df_ind, df_fam, id = NULL)
}
\arguments{
\item{evmat}{An event array}

\item{df_ind}{A dataframe describing individuals}

\item{df_fam}{A dataframe describing families}

\item{id}{Optional vector of IDs  to restrict to}
}
\description{
This function converts event array to Biograph
}
\examples{
\dontrun{
df_ind$bdate <- sample(seq(as.Date("1774-12-31"), as.Date("1874-12-31"), 100), nrow(df_ind))
df_fam <- data.frame(idf = c(0,unique(df_ind$momid[df_ind$momid>0])), fall = "C")
sample_kh(df_ind, df_fam)
}
}
\keyword{pedigree}

