% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matlist2array.R
\name{matlist2array}
\alias{matlist2array}
\title{Turns a list of sparse matrices into an array (or vice versa)}
\usage{
matlist2array(object)
}
\arguments{
\item{object}{Either a list of sparse matrices or an array with three dimensions}
}
\description{
This function returns an array of a given list of sparse matrices (or vice versa a list of sparse matrices for a given 3-dimensional array).
}
\examples{
\dontrun{
matlist2array(1000)
}
}
\keyword{kh}
\keyword{spouse}

