% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean_date.R
\name{mean_date}
\alias{mean_date}
\title{Mean date in intervals}
\usage{
mean_date(df = NULL, ldate = NULL, rdate = NULL)
}
\arguments{
\item{df}{A dataframe 1}

\item{ldate}{A date 1}

\item{rdate}{A date 2}
}
\description{
This function returns a dataframe being required to contruct pedigrees via the _kinship2::pedigree()_ function
}
\examples{
\dontrun{
#TBD
}
}
\keyword{import}
\keyword{kh}

