% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mem_check.R
\name{mem_check}
\alias{mem_check}
\title{Check for memory limits}
\usage{
mem_check(limit = 999, error_message = NULL, only_warn = FALSE)
}
\arguments{
\item{limit}{Numerical value for maximum memory usage allowed (Default is 999 MB)}

\item{error_message}{Error message to be returned, in case of exceeded memory limit.}

\item{only_warn}{Should instead only a warning be given? (Default is FALSE)}
}
\description{
This function tests whether the currently used memory
is above a given limit (Default to 999 MB) and either
gives out an error or the value of it's second argument.
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{memory}

