% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/others.R
\name{na2zero}
\alias{na2zero}
\title{Replace NAs with zeros}
\usage{
na2zero(x, zero = 0)
}
\arguments{
\item{x}{A R object containing zero to be replaced by NAs.}

\item{zero}{Replacement for missing values in x (Default is 0)}
}
\description{
This function replaces NAs with zeros.
}
\examples{
na2zero(c(1,2,NA,3,5))
}
\keyword{NA}

