% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_categories.R
\name{plot_categories}
\alias{plot_categories}
\title{A Barplots for Differences in Bionomial Probabilities}
\usage{
plot_categories(df, var = "var", status = "status", new_size = NA)
}
\arguments{
\item{df}{A dataframe containing status and factor variables}

\item{var}{Variable of interest}

\item{status}{Variable for status}

\item{new_size}{Change text size in labels}
}
\description{
This function plots estimated probabilities with 95 percent confidence intervals.
}
\examples{
plot_categories(data.frame(status = sample(c(0,1), 1000, replace=TRUE),
         group = factor(sample(c("A", "B", "C", "D"), 1000, replace=TRUE))), var="group")
}
\keyword{barplots}
\keyword{confidence}
\keyword{intervals}

