% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_kinmap.R
\name{plot_kinmap}
\alias{plot_kinmap}
\title{Plot a pedigree map}
\usage{
plot_kinmap(id = NULL, evdat = NULL, list_kin = NULL, list_geo = NULL,
  my_map = NULL, spit_results = FALSE, throw_plots = TRUE)
}
\arguments{
\item{id}{The individual ID of interests}

\item{evdat}{A specific date of interest}

\item{list_kin}{A nested list of dataframes providing the variables 'id', 'sex', 'relmom', 'reldad', 'age', and 'kinloc'}

\item{list_geo}{A list of geo stuff.. with geopositional data for locations}

\item{my_map}{A ggmap object}

\item{spit_results}{Spit it? (default is FALSE)}

\item{throw_plots}{Print'em? (default is TRUE)}
}
\description{
\code{ggmap} provided plot of a map showing the estimated locations of living pedigree members for a given ID at a given date. The size, shape and color of symbols represent for a specific individual's age, sex, and relatedness to an indiviudal x.
}
\examples{
\dontrun{
plot_kinmap(40, "1860-01-01", df_ind, relate, evmat, geo$gmap, geo$df)
}
}
\keyword{kh}
\keyword{pedigree}

