% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_val.R
\name{r_val}
\alias{r_val}
\title{Returns the relatedness coefficient between two individuals of a given pedigree}
\usage{
r_val(ped = NULL, x = NULL, y = NULL)
}
\arguments{
\item{ped}{Pedigree object for individuals 1 and 2}

\item{x}{ID of individual 1}

\item{y}{ID of individual 2}
}
\description{
This function returns the relatedness coefficient between two individuals of a given pedigree.
}
\examples{
exmpl_df <- get_exmpl_df()
ped <- with(exmpl_df, kinship2::pedigree(id, dadid, momid, sex, affected, status))
r_val(ped, exmpl_df$id[1], exmpl_df$id[2])
}
\keyword{kh}
\keyword{pedigree}
\keyword{relatedness}

