% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random_bio.R
\name{random_bio}
\alias{random_bio}
\title{Generate random biographies of reproductive event}
\usage{
random_bio(n, evmat_bak = FALSE, twinrate = 0.05)
}
\arguments{
\item{n}{Number of desired cases}

\item{evmat_bak}{Should twin births appear as multiple events at one date (Default FALSE)}

\item{twinrate}{Default to p = 0.05 (i.e. 5 percent twins)}
}
\description{
This function returns a generated event matrix
}
\examples{
random_bio(5)
}
\keyword{pedigree}

