% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recode_evloc.R
\name{recode_evloc}
\alias{recode_evloc}
\title{Recode locality labels from character strings to numeric (and vice versa)}
\usage{
recode_evloc(loc = NULL, evloc_list = list(name = c("CA", "CH", "CN", "CI",
  "EI", "FR", "GS", "GI", "GH", "GM", "HW", "HI", "JE", "LA", "LV", "LP", "LO",
  "MA", "MW", "NE", "PE", "PI", "RY", "SU", "TW", "UP", "UT", "VI", "WE", "WI",
  "WZ", "WO", "WY", "x.out", "x.unknown"), label = 1:35))
}
\arguments{
\item{loc}{Locality label}

\item{evloc_list}{A Locality label-number list}
}
\description{
Depending on chosen input format, this function return a vector containing the lavbels,
either numbers ord characters, each consistent
}
\examples{
recode_evloc(1:10)
recode_evloc("WY")
}
\keyword{kh}
\keyword{localitiy}

