% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/restrict2sample.R
\name{restrict2sample}
\alias{restrict2sample}
\title{Restrict data to a given sample size}
\usage{
restrict2sample(df_evn = NULL, df_ind = NULL, list_kin = NULL,
  array_mat = NULL, list_src = NULL, df_twn = NULL, param = NULL)
}
\arguments{
\item{df_evn}{A dataframe in long format holding event data}

\item{df_ind}{TBD}

\item{list_kin}{TBD}

\item{array_mat}{TBD}

\item{list_src}{TBD}

\item{df_twn}{TBD}
}
\description{
This function returns a dataframe being required to contruct pedigrees via the _kinship2::pedigree()_ function
}
\examples{
\dontrun{
kh_df <- get_df(df_ind, df_fam, get_fam=FALSE)
}
}
\keyword{kh}
\keyword{sample}
\keyword{size}

