% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_time.R
\name{run_time}
\alias{run_time}
\title{Run time since t_start}
\usage{
run_time(t_start = as.POSIXct(1, origin = "1981-12-13"),
  print_message = TRUE)
}
\arguments{
\item{t_start}{Time, for which time difference should be computed}

\item{print_message}{Set to FALSE, if to turn off messages and return a `difftime` instead.}
}
\description{
This function returns the time difference from `Sys.time()` to `t_start`, either as printed message (Default) or as `difftime`.
}
\examples{
\dontrun{
grap_ped(1067, df_ped)
}
}
\keyword{pedigree}

