% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_kh.R
\name{sample_kh}
\alias{sample_kh}
\title{Sample a Krummh\enc{ö}{oe}rn mother}
\usage{
sample_kh(df_ind, df_fam, n = 1)
}
\arguments{
\item{df_ind}{A dataframe}

\item{df_fam}{A dataframe containing}

\item{n}{Number of draws to perform. For n equal or larger the number of mothers in initial sample, the vector of all mother IDs is returned instead. Default is 1.}
}
\description{
This function returns random draws of mother IDs from a selection of families being classified as 'fully documented'.
}
\examples{
df_ind <- get_exmpl_df()
df_ind$bdate <- sample(seq(as.Date("1774-12-31"), as.Date("1874-12-31"), 100), nrow(df_ind))
df_fam <- data.frame(idf = c(0,unique(df_ind$momid[df_ind$momid>0])), fall = "C")
sample_kh(df_ind, df_fam)
}
\keyword{pedigree}

