% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/others.R
\name{trim}
\alias{trim}
\title{Remove white space from strings}
\usage{
trim(x)
}
\arguments{
\item{x}{A string}
}
\description{
This function removes blanks before and after a string expression (found via \href{http://stackoverflow.com/a/2261149}{SO, credit goes to f3lix}).
}
\examples{
trim("   Yo, wuzzup?     ")
}
\keyword{blanks}
\keyword{remove}
\keyword{string}

