% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unlist_df.R
\name{unlist_df}
\alias{unlist_df}
\title{Combine dataframes in a list by rows}
\usage{
unlist_df(dflist = NULL)
}
\arguments{
\item{dflist}{A list of dataframes}
}
\description{
This function returns the elements of a list of dataframes as one single dataframe (combined by rows, i.e. all dataframes must have the same number of columns).
}
\examples{
a_df <- data.frame(a=c(101,202), b=1:2, c=c("A", "B"))
b_df <- data.frame(a=c(303,404), b=3:4, c=c("C", "D"))
test <- list("a"=a_df, "b"=b_df)
unlist_df(test)
}
\keyword{kh33}
\keyword{spouse}

