% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valid_dates.R
\name{valid_dates}
\alias{valid_dates}
\title{Valid dates}
\usage{
valid_dates(id = NULL, evmat = NULL, names = FALSE)
}
\arguments{
\item{id}{An ID included in 'dimnames(evmat)'}

\item{evmat}{An event array (see '?get_evmat')}

\item{names}{Should output include the names of events? Default is false}
}
\description{
For a given ID, this function returns a vector of all event dates for which 'status' given in `evmat`is 1.
}
\examples{
\dontrun{
valid_dates(1570, kh_mat)
}
}
\keyword{distance}
\keyword{pedigree}
\keyword{spatial}

