% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/others.R
\name{zero2na}
\alias{zero2na}
\title{Replace zeros with NAs}
\usage{
zero2na(x, zero = 0, na = NA)
}
\arguments{
\item{x}{A R object containing zero to be replaced by NAs.}

\item{zero}{Value to be replaced (Default is 0)}

\item{na}{Replacement for value given in zero (Default is NA)}
}
\description{
This function replaces zeros with NAs.
}
\examples{
zero2na(c(1,2,0,3,5))
}
\keyword{NA}

